Extract ElementsElementHandler to separate class
